+++ pulsar-transaction/pom.xml
@@ -33,7 +33,6 @@
 
   <modules>
     <module>common</module>
-    <module>buffer</module>
     <module>coordinator</module>
   </modules>
